Bug 24281: Fix the list of types of visual materials
In the MARC21 XSLT files for results in both staff and OPAC, the
type of visual material for code "g" is "legal article" instead
of "game". This patch change the value to "game".

To test:
0. Set up:
    - have a MARC21 database
    - use default XSLT file for display of results :
        OPACXSLTResultsDisplay : default
        XSLTResultsDisplay : defaul
    - If needed, catalog a game. In the 008 tag editor :
        Type of Material : VM - Visual Materials
        33 - Type of visual material : g - Game
1. On both staff and OPAC, make a search which returns results
including visual materials of type game (the title catalogued
in previous step).
2. Note that in the list of results, the label "Type of visual
material" is followed by "legal article".
3. Apply patch.
4. Repeat step 1.
5. Note "legal article" has been replaced by "game".

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>